requirements.txt
alembic==1.7.5
astroid==2.9.1
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.27
platformdirs==2.4.1
protobuf==3.19.1
pylint==2.12.2
PyMySQL==1.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.29
toml==0.10.2
typed-ast==1.5.1
typing_extensions==4.0.1
visitor==0.1.3
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
